Add initial implementation of Compound Types
Currently, all compound types must be of the form ":foo" if the type is
declared to be of type "foo". This is because not requiring the colon
will take a lot of grammar editing and will make the parser much slower.
So for development purposes the colon exists, and it will be removed
later.

So far, only fields are implemented, and constructors are not, so there
is no way to create a compound type, only to access fields on them once
they exist.
Also, there is no way to assign to their fields yet.