Add native specifiers
These are modifiers which can be applied to a method, with the syntax:
native void foo() {...}
native "someName" void bar() {...}

They allow the programmer to specify an alternate name for the method,
which can be used in addition to the generated mangled name.

In the case that a string literal is not provided, the name of the
method is taken as the native name (e.g. "foo" in the first example
above).

Certain native names are forbidden. For example duplicate native names
are disallowed, as are names resembling an LLVM intrinsic function, and
also "malloc" and "free".