Add the parsing for the '?' prefix for nullability
Eventually, all types will be non-nullable by default, and it will be
possible to make any type nullable by prefixing it with a '?'.

This commit only adds the '?' to the language grammar and the parser,
and fixes all of the errors that having a 'nullable' field on every Type
introduces.