Finish adding nullable types
This adds support in the type checker and the code generator.

In the LLVM representation, most nullable types are represented as a
tuple of a boolean with the normal type. For example, a uint would be:
i32
whereas a ?uint would be:
{i1, i32}
The boolean (i1) value is 0 when the value represents null, and 1 when
it represents the second element of the tuple. This work similarly for
all other primitive types, tuple types, and compound types.
For array types (and eventually object types) the LLVM type does not
change, we use real null pointers instead.