Change static field access + array creation syntax
This modifies the grammar to make it much easier to see where a Type
ends and a static field access begins. The new operator is the double
colon '::', used as follows:
Type::staticField

This is required for qualified names to be used in types. Specifically,
in the case of an array of objects, the old grammar was ambiguous for
input:
compound foo { uint main() { []a.b.c.d = 5; } }
Here, we do not know where the type ends and the field access begins, so
we cannot construct an AST. In the new scheme, the syntax makes it
immediately obvious:
compound foo { uint main() { []a.b::c.d = 5; } }

There is a similar problem with array creations, because their syntax
has a trailing type. To solve this, we enforce that a field access
cannot be directly on an array creation, but must have brackets around
it.