Allow qualified names for types
This adds a package hierarchy to the main compiler and the Resolver,
which allows names to be resolved from the context of the root package
rather than just the current compilation unit.

Also, QName has been moved from a parse-time object to part of the AST.
This is because it is much easier to work with than a String[], and has
a LexicalPhrase associated with it.