Allow multiple source files to be compiled at once
This adds an ArgumentParser, which parses the command line arguments,
and currently looks for "-d <dir>" to decide which directory to output
the bitcode files to. The bitcode files are added in a directory
structure based on the package hierarchy under the output directory.

To do this, the resolver needed to be split into 3 stages:
* add type definitions to their packages
* resolve the top level types
* resolve everything else (expressions and statements)

Also, the code generator needed to be changed to look up functions in
such a way that it could generate declarations for functions which are
not in the same module.