Add null literal expressions, and type propagation
Type propagation is a system that propagates certain types from
super-expressions into sub-expressions. It runs after the type checker
has checked the types of everything, and because of this, running it can
never result in a ConceptualException.

The main use for it is to deduce the type of the null literal in
something like:
(test ? a : null)?.foo();
Here, it is impossible to know the type of the null literal until after
both of the sub-expressions of the inline-if have been type-checked.
Therefore, after the TypeChecker has run, the TypePropagator has to find
the InlineIfExpression and propagate its now-known type down to the
NullLiteralExpression.

The reason this has to be done is that without it the CodeGenerator
cannot know what type of null to produce. A null literal can be a simple
null pointer for an ArrayType, but e.g. for a PrimitiveType it must be a
tuple of a boolean and a primitive value.

It is not an error for the type propagator to be unable to remove all of
the NullTypes from an AST. In these cases, we do not care about the
actual type of the null literal, because it is never used. For example:
a, _ = 5, null;
We therefore allow the code generator to translate a NullType into an
empty LLVM tuple.