Add the null-coalescing operator '?:'
This operator takes two operands, and is right-associative, with the
same precedence as the inline-if operator.

The first operand must be nullable, and the second may or may not be.
Crucially, the non-nullable version of the first operand must have a
common super-type with the second operand, which becomes the result type
of the NullCoalescingExpression.