Add metadata generation
This adds metadata to each generated module, containing information
about the top level of a given compound definition, including fields,
constructors, methods, and their types and parameters.

This information is intended to be read back in by the compiler in order
to allow other files to be compiled against a bitcode file without
recompiling it from a source file in the process.