Keep output position while lazily adding methods
This could cause a rare bug if a method was added for the first time
during generation of a call to that method.
If a native method existed, it would be created and code would be added
to it to call the non-native version of the method, which would
reposition the builder to the end of the native method, where it would
remain for any subsequent code generation in the caller.