Add final variables
Currently, only local variables, parameters, and non-static fields may
be final. Static final fields will be supported later on, but as yet
there is no way to initialise them.

The final modifier only guarantees that a variable can only be assigned
once, not that its value can never change. There is a subtle difference
that will not be visible until static final variables are allowed:
before a global (static) field is initialised, its value will default to
null or zero, and can be read as such if another global variable's
initialiser reads it before it is initialised.

The final modifier changes nothing about the generated code, it only
places restrictions on when variables can and cannot be modified.