Ensure declarations happen; prevent redeclarations
Previously, a variable declaration would only happen if there was not a
field with the same name, which was wrong.
Also, a declaration in which all of the variables had previously been
declared would work fine, despite the fact that they should not be
redeclared.
This commit fixes both of these problems.