Split Comparisons into Equality and Relational
This not only splits them into two different types, but allows Equality
comparisons on a much broader set of types than Relational comparisons
(which can only be used on integers and floating point values).

'==' and '!=' can now be used to check whether something is null, or
check whether two values of the same type are equal to each other.
If two values are not the same type, but can be converted to a common
supertype before the comparison, then the comparison can be performed
with that type.