Add ability to load metadata from bitcode files
This large patch adds a MetadataLoader, which loads the type of metadata
outputted by MetadataGenerator into CompoundDefinitions.

This commit currently depends on some C bindings which only exist in an
unreleased version of LLVM, so to build the compiler you now need to
build LLVM from source. This requires a checkout of LLVM with an SVN
revision of 164247 or later (or, when it is released, LLVM 3.2).

To use this to compile a source file against an existing bitcode file,
that bitcode file must currently be in the place where the compiler
would put it if it was compiling it with the same -d (or --output-dir)
option. In effect, the only imported directory is currently the output
directory. This should change soon.

If you compile something against an old version of another bitcode file,
and then prior to linking you switch to a newer version which has had
binary incompatible changes made to it, then the behaviour is undefined,
and may break randomly (as you might expect for binary incompatible
changes).