Clean up native types
Native types now have two distinct native representations: standard and
temporary.

The standard type representation is used for passing values into
function calls, returning values from functions, storing values in
arrays and object and compound types, etc. The temporary type
representation is used for storing things internally to a function.

The main differance between the two representations is that compound
(and partially tuple) types are represented differently: in the standard
representation, compound types are tupled with an i1 to indicate their
nullity, whereas in the temporary representation they are made into
pointers to the values. This allows methods to be called on compound
types which are allocated on the stack or on the heap.

Tuple types have a slightly different temporary representation, which is
similar to their standard representation, but their subtypes are stored
in the temporary representation instead of the standard representation.

This commit adds TypeHelper, which simplifies the conversion that
CodeGenerator has to do, in favour of just making
typeHelper.convert...() calls wherever necessary.