Add initialisers
Initialisers can be defined by placing blocks directly in a type
definition, with an optional static modifier. For example:
static { x = doSomething(); }
{ o = this.foo(); }

Initialisers for single fields can also be defined by treating a field
declaration as an assignment:
static Foo foo = new Foo();
?Bar bar = null;

Normal initialisers are run at the beginning of a constructor. When
inheritance is added, they will be run after the superclass's
constructor.
Static initialisers are run when the program is loaded, using the
'llvm.global_ctors' global variable, which stores an array of functions
to be run before main().