Fix crash when the first initialiser isn't a Field
Before this commit, when the first initialiser wasn't a field, the code
generator would try to store all of its variables at the start of the
entry block, before there were any instructions in the entry block.
Because it was trying to position the builder before a null first
instruction, LLVM would crash.
The new LLVMPositionBuilderAtStart() fixes this easily.