Change the Type syntax to remove the preceding ':'
This required a large number of changes, most of them being grammar
hacks.
The thing which actually caused a problem was a conflict between
types and assignees. The following two lines demonstrate this problem:
A (b).f = g();
a(b).f = g();
These parse to the same types of tokens, but one normally represents a
declaration of something of type 'A', which turns out to be a field
access. The other normally represents a call to a function called 'a'.

Since it is generally impossible to tell which case the user expects,
this is impossible to solve. Therefore, one of the cases had to be made
invalid. Since declaring the type of a field which already exists is not
a common case, and could make for more confusing code, the assignees for
declarations had to be limited to actual declarations (simple names) and
blank assignees (underscores).