Add Linking
This allows the user to specify a final output file, which will store a
linked version of everything in all of the compiled modules, as well as
all linked-in bitcode files specified with '--link'.
Compiled modules can still be outputted separately using '--output-dir',
but using '--output' will now cause a linked version to be written too.

This also adds the ability to specify bitcode files with '--import',
which means that they will be loaded into the package structure, but
their contents will not be linked into the final output file.