Allow building the runtime on other platforms
Previously, the runtime depended on non-standard features of the C
library on linux, specifically: the linker names for stdin, stdout, and
stderr.
Now, these are used through the functions in plinth-lib/core/c/io.c,
which is platform independent, so it should work on any platform that
supports the standard C library.
The downside is that there is now a C file, which must be compiled to
bitcode using clang -emit-llvm in order for programs to build.