Try harder to resolve method call ambiguities
Previously, a call of foo(1) would be ambiguous if two overloaded
versions of foo() took either a ubyte or a ushort. This would cause
compiler errors even though we know that the argument type is ubyte.

Now, instead of this, we prefer to call methods which have parameters
which are equivalent to the provided arguments. For example, foo(1)
above would resolve to the ubyte overload of foo, since 1 is a ubyte.

Also, if an equivalent parameter list cannot be found, we try to find
one which is equivalent after making arguments nullable wherever
necessary.