Add checking and codegen for delegate constructors
This changes how constructors are called, since before a compound
constructor would create its own value. Now, all constructors take an
allocated piece of memory for 'this' and initialise it, before returning
void (rather than the initialised value).

This change to the layout of constructor functions allows all
constructors to be called as delegates.