Add immutability
This is a very large set of changes whose main purpose is to add
immutability. This is in the form of a new 'immutable' modifier for
constructors and methods, and a change to array, function, and named
types to allow them to be immutable, using a '#' prefix.

Immutable array and named types cannot have any of their elements
modified (even recursively, so e.g. changing a field of an array element
is also disallowed), while non-immutable functions cannot be called from
an immutable context.

This also changes the order of some passes. Previously, there was no
constraint on which order the control flow checker and type checker ran
in, and the control flow checker ran first. Now, the type checker must
be run before the control flow checker, so that it can propagate the
immutability of expressions for the control flow checker to check.