Add the 'mutable' field modifier
This modifier can only be applied to fields. A mutable field can be
modified from an immutable context, by either assigning a different
value to it, or by changing some of its internal state.
The final modifier and the immutability of the type take precedence over
this modifier, so a 'mutable #Object' would only allow the field to be
reassigned, and a 'final mutable #Object' is not allowed, since the
mutable modifier no longer means anything.