Add the object type, and ensure it has toString()
This adds the type 'object', which is the supertype of all other types
(once we consider it with nullability and immutability).
To convert to object, a heap allocation must be performed for any value
stored on the stack. Values stored on the heap do not require a heap
allocation (although the array representation does not yet take
advantage of this).
To convert back to a normal type, the value stored in the object can
simply be extracted.

This commit also adds toString() methods to all types, since they were
required for object's virtual function table to work correctly.
All types get their own object-style virtual function tables, which can
be directly stored into new objects which are being converted from them.
All NamedTypes automatically get built-in toString() methods, if they
are not explicitly overridden. We ensure that class definitions have
these built-in object methods at the start of the VFT of the first
object in their native representation.