Add selfish constructors; fixes escape analysis
Previously, there was a bug in the escape analysis, which allowed 'this'
to escape from a constructor before the object was fully initialised.

This was because a super-constructor could always use 'this' after it
had initialised its own variables, and it would be called before the
subclass's initialiser had been run.

To solve this, a new 'selfish' modifier has been added to constructors,
which prevents them from being used as a super-constructor, but allows
them to use 'this' and call instance methods as before.

Default non-selfish constructors are now forbidden from using 'this', or
allowing it to escape in any way.