Fix an error with using types loaded from bitcode
Previously, a type loaded from bitcode only had the first
top-level-types resolution pass run on it, which meant that it never got
a linearisation, and so resolving its members would cause a
NullPointerException.

Now, the PassManager brings each imported bitcode file up to speed as
they are loaded, so that they can instantly be used to the same extent
as any bitcode file specified on the command line.