Allow multiple errors to be output in a single run
This adds a CoalescedConceptualException, which stores multiple
ConceptualExceptions until they all need to be thrown.

This patch required major changes to several passes, to allow enough
Exception caching to be useful, without creating a situation that could
cause NullPointerExceptions. To ensure this, we stop running passes as
soon as a single pass throws an error. We also make sure that passes do
not continue once an unrecoverable error has been detected (e.g. trying
to resolve a function call when its arguments do not type-check, or
trying to type-check something which has failed to resolve).

Because of this, TypeDefinitions loaded from metadata must be discarded
if they cannot immediately be brought up to speed with the current pass,
or they could cause NullPointerExceptions later on in resolution when
something else tries to resolve to the same TypeDefinition.