Make function immutability cast-away-able
This makes function immutability a property of the value, not of the
reference. So it will be possible to cast an immutable function value
which was stored as not-immutable (or as an object) back into an
immutable function value.

Eventually, a run-time exception will be thrown when someone tries to
cast a not-immutable value to immutable.