Add a prototype for zero-cost exception handling
The way this works is: each call site uses an invoke, which can branch
to a landingpad in the case of an exception. The landingpad has a
personality function, which determines whether the unwinder should stop
at that landingpad while unwinding a given exception. This personality
function parses some debugging data generated by LLVM in order to check
the types of exceptions that the current landingpad supports against the
actual exception object's type. This debugging data may be target
specific, which means the personality function must also be target
specific. So far, the only personality function is the one for x86, and
it has only been tested on x86_64.

Some relevant links detailing the implementation of this are:
http://llvm.org/docs/ExceptionHandling.html
http://mentorembedded.github.com/cxx-abi/abi-eh.html
http://mentorembedded.github.com/cxx-abi/exceptions.pdf