Don't just continue if we run out of memory: abort
This may be changed to an error-throwing implementation later on. In
order to do that, an emergency heap and a backup allocator need to be
written in order to create the exception and store the whole stack trace
inside it.
For now, it just prints a message to stderr, and aborts.