Add CastError, and throw it if a cast fails
This also adds Throwable and Error, which are core parts of the
exception handling hierarchy. Exception will be added once it can be
thrown.

Another addition is a function which is called by the runtime whenever
an uncaught exception is encountered:
Throwable::printUncaughtException()
This function is a native upcall which prints the exception, and is
called from the exception handling runtime at
plinth-lib/core/c/exceptions.c