Fix TypeChecker::findCommonSuperType() wildcards
This method tries to generate wildcard type arguments in order to find
the lowest common super-type possible for a NamedType.
When it finds two NamedTypes for the same TypeDefinition which have
different type arguments, it will combine the type arguments into a
single wildcard type argument instead of falling back to the object
type.
The generated wildcard will be of the form '? extends X' where X is the
common super-type of all of the type arguments for this type parameter,
or it will default to just '?' if there is no common super-type.
Since it uses findCommonSuperType() to find X, it can choose X=double if
the type arguments are for example uint and double.
The problem with this is that at runtime, a conversion from Foo<uint> to
Foo<? extends double> will fail, because the runtime type checker checks
that the wildcard matches before allowing the conversion, and it has
slightly different rules about conversion, which say that primitive
types cannot be converted between each other.
(The reason for this is that for conversions between primitive types to
be possible, the source and destination types must both be known at
compile time.)
This mismatch in rules means that there was previously no way to detect
this type of scenario at compile time. Therefore, a new method was added
to Type called canRuntimeAssign(), which is like canAssign() but with
the runtime's rules.
In findCommonSuperType(), we now explicitly try to detect this scenario
when generating wildcard super-types, and if a canRuntimeAssign() check
fails, we just use '?' as the super-type.